webpack support and move tests into fixtures directory
