--- tests/data/cddl/precedence01.cddl
@@ -1,3 +1,0 @@
-t = [group1]
-group1 = (a / b // c / d)
-a = 1 b = 2 c = 3 d = 4
