--- tests/data/cddl/reputon.cddl
@@ -1,38 +1,0 @@
-reputation-object = {
-  reputation-context,
-  reputon-list
-}
-
-reputation-context = (
-  application: text
-)
-
-reputon-list = (
-  reputons: reputon-array
-)
-
-reputon-array = [* reputon]
-
-reputon = {
-  rater-value,
-  assertion-value,
-  rated-value,
-  rating-value,
-  ? conf-value,
-  ? normal-value,
-  ? sample-value,
-  ? gen-value,
-  ? expire-value,
-  * ext-value,
-}
-
-rater-value = ( rater: text )
-assertion-value = ( assertion: text )
-rated-value = ( rated: text )
-rating-value = ( rating: float16 )
-conf-value = ( confidence: float16 )
-normal-value = ( normal-rating: float16 )
-sample-value = ( sample-size: uint )
-gen-value = ( generated: uint )
-expire-value = ( expires: uint )
-ext-value = ( text => any )
