--- tests/data/cddl/socketplug.cddl
@@ -1,24 +1,0 @@
-PersonalData = {
-  ? displayName: tstr,
-  NameComponents,
-  ? age: uint,
-  * $$personaldata-extensions
-}
-
-NameComponents = (
-  ? firstName: tstr,
-  ? familyName: tstr,
-)
-
-; The above already works as is.
-; But then, we can add later:
-
-$$personaldata-extensions //= (
-  favorite-salsa: tstr,
-)
-
-; and again, somewhere else:
-
-$$personaldata-extensions //= (
-  shoesize: uint,
-)
