+++ tests/fixtures/cddl/byron.cddl
@@ -9,7 +9,7 @@ ebblock = [ "header": ebbhead, "body": [ + stakeholderid ], extra: [ attributes 
 u8 = uint .lt 256
 u16 = uint .lt 65536
 u32 = uint
-u64 = uint 
+u64 = uint
 
 ; Basic Cardano Types
 
@@ -26,12 +26,12 @@ epochid = u64
 slotid = [ epoch: epochid, slot: u64 ]
 
 pubkey = bytes
-signature = bytes 
+signature = bytes
 
 ; Attributes - at the moment we do not bother deserialising these, since they
 ; don't contain anything
 
-attributes = { * any => any } 
+attributes = { * any => any }
 
 ; Addresses
 
@@ -43,7 +43,7 @@ addrtype = &( "PubKey": 0, "Script": 1, "Redeem": 2 ) / (u64 .gt 2)
 
 addrattr = { ? 0: addrdistr, ? 1: bytes }
 
-address = [ #6.24(bytes .cbor ([ addressid, addrattr, addrtype ])), u64 ] 
+address = [ #6.24(bytes .cbor ([ addressid, addrattr, addrtype ])), u64 ]
 
 ; Transactions
 
@@ -58,7 +58,7 @@ txproof = [ u32, hash, hash ]
 twit = [ 0, #6.24(bytes .cbor ([ pubkey, signature ])) ]
 	/ [ 1, #6.24(bytes .cbor ([ [ u16, bytes ], [ u16, bytes ] ])) ]
 	/ [ 2, #6.24(bytes .cbor ([ pubkey, signature ])) ]
-	/ [ u8 .gt 2, encoded-cbor ] 
+	/ [ u8 .gt 2, encoded-cbor ]
 
 ; Shared Seed Computation
 
@@ -104,7 +104,7 @@ ssc = [ 0, ssccomms, ssccerts ]
 sscproof = [ 0, hash, hash ]
 	/ [ 1, hash, hash ]
 	/ [ 2, hash, hash ]
-	/ [ 3, hash ] 
+	/ [ 3, hash ]
 
 ; Delegation
 
@@ -124,7 +124,7 @@ lwdlg = [
 	certificate: signature
 ]
 
-lwdlgsig = [ lwdlg, signature ] 
+lwdlgsig = [ lwdlg, signature ]
 
 ; Updates
 
@@ -173,7 +173,7 @@ upvote = [
 	"signature": signature
 ]
 
-up = [ "proposal": [ ? upprop ], votes: [ * upvote ] ] 
+up = [ "proposal": [ ? upprop ], votes: [ * upvote ] ]
 
 ; Blocks
 
@@ -217,7 +217,7 @@ blockbody = [
 	"sscPayload": ssc,
 	"dlgPayload": [ * dlg ],
 	"updPayload": up
-] 
+]
 
 ; Epoch Boundary Blocks
