+++ tests/fixtures/cddl/precedence01.cddl
@@ -1,0 +1,8 @@
+t = [ group1 ]
+
+group1 = ( a / b // c / d )
+
+a = 1
+b = 2
+c = 3
+d = 4
